Custom Radio buttons implementation alleviating Lift's radio buttons pain
Radio button handling in Lift is horribly broken if you're doing anything remotely custom with the radios:
 - They replace the entire element removing all attributes you might select on for styling and breaking the
   accessibility of labels,
- They add hidden fields to the markup, which breaks the '+' CSS selector.
On a page where the labels are actually the only thing showing and we depend on relative positioning for things to display and behave correctly, both of those are deal breakers. This code is a custom implementation for Lift radios that should alleviate all the pain.